Prompt_Optimizer/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
google-generativeai==0.3.1
pydantic==2.4.2
python-dotenv==1.0.0 